multi-route-chain-app/requirements.txt
attrs==23.1.0
autopep8==2.0.4
aws-cdk-lib==2.91.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
cattrs==23.1.2
cdk-nag==2.27.209
constructs==10.3.0
importlib-resources==6.1.1
jsii==1.92.0
publication==0.0.3
pycodestyle==2.11.1
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.8.0
